+++ konva.js
@@ -3,7 +3,7 @@
  * Konva JavaScript Framework v0.11.0
  * http://konvajs.github.io/
  * Licensed under the MIT or GPL Version 2 licenses.
- * Date: Sat Nov 21 2015
+ * Date: Sun Nov 22 2015
  *
  * Original work Copyright (C) 2011 - 2013 by Eric Rowell (KineticJS)
  * Modified work Copyright (C) 2014 - 2015 by Anton Lavrenov (Konva)
@@ -799,6 +799,13 @@ var Konva = {};
                 }
             }
             return names.length > 0;
+        },
+        isValidSelector: function(selector) {
+            if (typeof selector !== 'string') {
+                return false;
+            }
+            var firstChar = selector[0];
+            return firstChar === '#' || firstChar === '.' || firstChar === firstChar.toUpperCase();
         },
         createCanvasElement: function() {
             var canvas = Konva.document.createElement('canvas');
@@ -2658,8 +2665,17 @@ var Konva = {};
          *   var oldVal = evt.oldVal;
          *   var newVal = evt.newVal;
          * });
+         *
+         * // also event delegation works
+         * layer.on('click', 'Group', function(evt) {
+         *   var shape = evt.target;
+         *   var group = evtn.currentTarger;
+         * });
          */
         on: function(evtStr, handler) {
+            if (arguments.length === 3) {
+                return this._delegate.apply(this, arguments);
+            }
             var events = evtStr.split(SPACE),
                 len = events.length,
                 n, event, parts, baseEvent, name;
@@ -2756,6 +2772,17 @@ var Konva = {};
         },
         removeEventListener: function(type) {
             this.off(type);
+        },
+        _delegate: function(event, selector, handler) {
+            var stopNode = this;
+            this.on(event, function(evt) {
+                var targets = evt.target._findMatchers(selector, stopNode);
+                for(var i = 0; i < targets.length; i++) {
+                    evt = Konva.Util.cloneObject(evt);
+                    evt.currentTarget = targets[i];
+                    handler.call(targets[i], evt);
+                }
+            });
         },
         /**
          * remove self from parent, but don't destroy
@@ -3408,6 +3435,49 @@ var Konva = {};
         getParent: function() {
             return this.parent;
         },
+        _findMatchers: function(selector, stopNode) {
+            var res = [];
+            if (this._isMatch(selector)) {
+                res.push(this);
+            }
+            var parent = this.parent;
+            if (!parent) {
+                return res;
+            }
+            if (parent === stopNode) {
+                return res;
+            }
+            return res.concat(parent._findMatchers(selector, stopNode));
+        },
+        _isMatch: function(selector) {
+            var selectorArr = selector.replace(/ /g, '').split(','),
+                len = selectorArr.length,
+                n, sel;
+
+            for (n = 0; n < len; n++) {
+                sel = selectorArr[n];
+                if (!Konva.Util.isValidSelector(sel)) {
+                    Konva.Util.warn('Selector "' + sel + '" is invalid. Allowed selectors examples are "#foo", ".bar" or "Group".');
+                    Konva.Util.warn('If you have a custom shape with such className, please change it to start with upper letter like "Triangle".');
+                    Konva.Util.warn('Konva is awesome, right?');
+                }
+                // id selector
+                if(sel.charAt(0) === '#') {
+                    if (this.id() === sel.slice(1)) {
+                        return true;
+                    }
+                }
+                // name selector
+                else if(sel.charAt(0) === '.') {
+                    if (this.hasName(sel.slice(1))) {
+                        return true;
+                    }
+                } else if (this._get(sel).length !== 0) {
+                    return true;
+                }
+            }
+            return false;
+        },
         /**
          * get layer ancestor
          * @method
@@ -3462,12 +3532,14 @@ var Konva = {};
          */
         fire: function(eventType, evt, bubble) {
             // bubble
+            evt = Konva.Util.cloneObject(evt || {});
+            evt.currentTarget = this;
             if (bubble) {
-                this._fireAndBubble(eventType, evt || {});
+                this._fireAndBubble(eventType, evt);
             }
             // no bubble
             else {
-                this._fire(eventType, evt || {});
+                this._fire(eventType, evt);
             }
             return this;
         },
@@ -6377,14 +6449,6 @@ var Konva = {};
 
 (function() {
     'use strict';
-
-    function isValidSelector(selector) {
-        if (typeof selector !== 'string') {
-            return false;
-        }
-        var firstChar = selector[0];
-        return firstChar === '#' || firstChar === '.' || firstChar === firstChar.toUpperCase();
-    }
     /**
      * Container constructor.&nbsp; Containers are used to contain nodes or other containers
      * @constructor
@@ -6583,7 +6647,7 @@ var Konva = {};
 
             for (n = 0; n < len; n++) {
                 sel = selectorArr[n];
-                if (!isValidSelector(sel)) {
+                if (!Konva.Util.isValidSelector(sel)) {
                     Konva.Util.warn('Selector "' + sel + '" is invalid. Allowed selectors examples are "#foo", ".bar" or "Group".');
                     Konva.Util.warn('If you have a custom shape with such className, please change it to start with upper letter like "Triangle".');
                     Konva.Util.warn('Konva is awesome, right?');